pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "logic-flow-analysis"
version = "3.1.0"
description = "Windows Kernel Driver Logic Flow Analysis Tool"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Security Research Tools"}
]
keywords = [
    "ida pro",
    "ghidra", 
    "reverse engineering",
    "windows kernel",
    "driver analysis",
    "security research",
    "binary diffing"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Disassemblers",
]

dependencies = [
    "PyQt6>=6.5.0",
    "PyQt6-Qt6>=6.5.0",
    "psutil>=5.9.0",
    "pefile>=2023.2.7",
]

[project.optional-dependencies]
gui = [
    "PyQt6-WebEngine>=6.5.0",
]
analysis = [
    "py-tlsh>=4.5.0",
    "ssdeep>=3.4",
]
symbolic = [
    "angr>=9.2.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]
all = [
    "logic-flow-analysis[gui,analysis,dev]",
]

[project.scripts]
logic-flow = "logic_flow.main:main"
logic-flow-gui = "run_gui:main"

[project.urls]
"Homepage" = "https://github.com/your-repo/logic-flow-analysis"
"Bug Tracker" = "https://github.com/your-repo/logic-flow-analysis/issues"
"Documentation" = "https://github.com/your-repo/logic-flow-analysis#readme"

[tool.setuptools]
packages = ["logic_flow", "logic_flow.core", "logic_flow.gui", "logic_flow.utils"]
include-package-data = true

[tool.setuptools.package-data]
"logic_flow.core" = ["ida_analysis_script.py"]
"logic_flow.gui" = ["assets/*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
